+++ WebSite/src/Home.js
@@ -1,31 +1,66 @@
 import React, { Component } from "react";
+import axios from 'axios';
+import jwt from 'jsonwebtoken';
 
 class Home extends Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      universal: null,
+      lastRound: null
+    }
+    axios({
+
+    })
+  }
+  getHighScores
+
   render() {
     return (
       <div>
-        <h2>HIGH SCORES</h2>
+        <div className="boxwrapper">
+          <h1 className="box">Universal ratings (top 5)</h1>
+          <h1 className="box">Last round ratings (top 5)</h1>
+        </div>
         <div className="boxwrapper">
           <div className="box whitebg">
-            <p>Cras facilisis urna ornare ex volutpat, et
-        convallis eat elementum. Ut aliquam, ipsum vitae
-        gravida suscipit, metus dui bibendum est, eget rhoncus nibh
-        metus nec massa. Maecenas hendrerit laoreet augue
-        nec molestie. Cum sociis natoque penatibus et magnis
-        dis parturient montes, nascetur ridiculus mus.</p>
+            <p>Loading high scores...</p>
           </div>
           <div className="box whitebg">
-            <p>Duis a turpis sed lacus dapibus elementum sed eu lectus.</p>
+            <p>Loading high scores...</p>
           </div>
-          <div className="box whitebg">
-            <p>Maecenas hendrerit laoreet augue
-        nec molestie. Cum sociis natoque penatibus et magnis
-        dis parturient montes, nascetur ridiculus mus.</p>
-          </div>
+        </div>
+        <div className="boxwrapper">
+          <p className="box">Your rank: </p>
+          <p className="box">Your rating: </p>
+          <p className="box">Your rank: </p>
+          <p className="box">Your rating: </p>
         </div>
       </div>
     );
   }
+}
+
+function getHighScores() {
+  axios({
+    method: 'post',
+    url: 'http://maltebp.dk:10000/getHighScores',
+    headers: {
+      'Content-Type': 'application/json',
+      'Access-Control-Allow-Origin': 'http://maltebp.dk:10000/getHighScores',
+      'Accept': 'application/json',
+      'Authorization': localStorage.getItem("dftoken")
+    },
+    withCredentials: true
+  })
+    .then(response => {
+      
+      //this.handleState(res);
+    })
+    .catch(error => {
+      console.log("axios download results error", error);
+    });
+  //event.preventDefault();
 }
 
 export default Home;
