+++ WebSite/src/index.css
@@ -4,17 +4,15 @@
 }
 
 a {
-  color:white;
+  color: white;
   text-decoration: underline overline;
 }
 
-
-body{
+body {
   background-image: url("images/df.jpg");
   background-color: #cccccc;
   color: #fff;
-  font-family: 'Pixellari', 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+  font-family: 'Pixellari', 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
   margin: 0 auto;
   max-width: 1000px;
 }
@@ -29,60 +27,56 @@ h1.header {
 }
 
 ul.header li {
-  display:inline;
+  display: inline;
   list-style-type: none;
-  margin:0;
+  margin: 0;
 }
 
 ul.header {
   text-align: center;
-  padding:0;
+  padding: 0;
 }
 
 ul.header li a {
-  color:#fff;
+  color: #fff;
   font-weight: bold;
   text-decoration: none;
-  padding:20px;
-  display:inline-block;
-  border-top:2px solid transparent;
-  border-bottom:2px solid transparent;
+  padding: 20px;
+  display: inline-block;
+  border-top: 2px solid transparent;
+  border-bottom: 2px solid transparent;
 }
 
 .active {
-  border-top:2px solid white !important;
-  border-bottom:2px solid white !important;
+  border-top: 2px solid white !important;
+  border-bottom: 2px solid white !important;
 }
 
-.content h1, .content h2, .content h3, .box{
-  /*margin:1%; /*To reflect box margin*/
-  text-align: center;
+.box {
+  margin: 1%;
 }
 
-.box{
-  margin:1%;
-}
-
-.boxwrapper{
+.boxwrapper {
   display: -webkit-flex;
   display: -ms-flexbox;
   overflow: hidden;
   font-family: monospace;
 }
 
-.whitebg{
-  background:white;
-  color:black;
+.whitebg {
+  background: white;
+  padding: 20px;
+  color: black;
 }
 
-.box{
+.box {
   display: table-cell;
-  padding: 20px;
   box-sizing: border-box;
-  margin-bottom:20px;
+  margin-bottom: 20px;
 }
 
 /*Smaller screens*/
+
 @media only screen and (max-width: 600px) {
   .box, .box3, ul.header li {
     display: block;
@@ -91,23 +85,23 @@ ul.header li a {
 }
 
 /*Bigger screens*/
+
 @media only screen and (min-width: 600px) {
-.boxwrapper{
-display:flex;
-}
+  .boxwrapper {
+    display: flex;
+  }
   .box, .boxwrapper {
-   flex:1;
+    flex: 1;
   }
 }
 
-
 /*Style the web form*/
 
-form label, form input{
-  display:block;
-  width:100%;
+form label, form input {
+  display: block;
+  width: 100%;
   box-sizing: border-box;
-  color:white;
+  color: white;
   background-color: transparent;
   text-align: center;
   border: none;
@@ -117,19 +111,19 @@ form label, form input{
   font-size: 1.3em;
 }
 
-form button{
-  color:black;
-  width:100%;
+button, form button {
+  color: black;
+  width: 100%;
   background-color: #ffffff;
-  border:none;
-  padding:20px;
-  margin-top:10px;
-  margin-bottom:20px;
+  border: none;
+  padding: 20px;
+  margin-top: 10px;
+  margin-bottom: 20px;
   font-family: 'Courier New', Courier, monospace;
 }
 
-  form {
-    text-align: center;
-    margin: 0 auto;
-    max-width: 320px;
-  }
+form {
+  text-align: center;
+  margin: 0 auto;
+  max-width: 320px;
+}
